+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/PmdIncrementalTest.groovy
@@ -1,0 +1,45 @@
+package ru.vyarus.gradle.plugin.quality
+
+
+import org.gradle.api.Project
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 06.11.2019
+ */
+class PmdIncrementalTest extends AbstractTest {
+
+    def "Check not enabled by default"() {
+
+        when: "apply plugin"
+        file('src/main/java').mkdirs()
+
+        Project project = project {
+            apply plugin: 'java'
+            apply plugin: 'ru.vyarus.quality'
+
+            quality.pmdIncremental = false
+        }
+
+        then: "value applied to pmd"
+        project.pmd.incrementalAnalysis.get() == false
+
+    }
+
+    def "Check incremental value applied"() {
+
+        when: "apply plugin"
+        file('src/main/java').mkdirs()
+
+        Project project = project {
+            apply plugin: 'java'
+            apply plugin: 'ru.vyarus.quality'
+
+            quality.pmdIncremental = true
+        }
+
+        then: "value applied to pmd"
+        project.pmd.incrementalAnalysis.get() == true
+
+    }
+}